Obsolete DisposalTests by ensuring all scenarios are covered by LifecycleTests.
