--- src/Fixie.Tests/Lifecycle/DisposalTests.cs
@@ -1,44 +1,0 @@
-﻿namespace Fixie.Tests.Lifecycle
-{
-    using static System.Environment;
-
-    public class DisposalTests : BaseLifecycleTests
-    {
-        public void ShouldFailCaseWhenConstructingPerCaseAndDisposeThrows()
-        {
-            FailDuring("Dispose");
-
-            Convention.ClassExecution
-                      .CreateInstancePerCase();
-
-            var output = Run();
-
-            output.ShouldHaveResults(
-                "SampleTestClass.Pass failed: 'Dispose' failed!",
-                "SampleTestClass.Fail failed: 'Fail' failed!" + NewLine +
-                "    Secondary Failure: 'Dispose' failed!");
-
-            output.ShouldHaveLifecycle(
-                ".ctor", "Pass", "Dispose",
-                ".ctor", "Fail", "Dispose");
-        }
-
-        public void ShouldFailAllCasesWhenConstructingPerClassAndDisposeThrows()
-        {
-            FailDuring("Dispose");
-
-            Convention.ClassExecution
-                      .CreateInstancePerClass();
-
-            var output = Run();
-
-            output.ShouldHaveResults(
-                "SampleTestClass.Pass failed: 'Dispose' failed!",
-                "SampleTestClass.Fail failed: 'Fail' failed!" + NewLine +
-                "    Secondary Failure: 'Dispose' failed!");
-
-            output.ShouldHaveLifecycle(
-                ".ctor", "Pass", "Fail", "Dispose");
-        }
-    }
-}
